Explicitly set compute softmax in float32 (#135)
The [Keras mixed precision guide](https://www.tensorflow.org/guide/keras/mixed_precision) recommends computing the sofmax in float32 for better numerical stability. If I recall correctly this was done implicitely in TensorFlow 2.0 since the softmax op was [blacklisted in the mixed precision graph rewrite](https://github.com/tensorflow/tensorflow/blob/v2.0.0/tensorflow/core/grappler/optimizers/auto_mixed_precision_lists.h#L146-L174). As of TensorFlow 2.1.0 this is no longer the case so we should follow the recommendation of the guide and explicitly set the data type.

+++ larq_zoo/literature/binary_alex_net.py
@@ -79,7 +79,7 @@ class BinaryAlexNetFactory(ModelFactory):
             out = self.dense_block(out, units=4096)
             out = self.dense_block(out, units=4096)
             out = self.dense_block(out, self.num_classes)
-            out = tf.keras.layers.Activation("softmax")(out)
+            out = tf.keras.layers.Activation("softmax", dtype="float32")(out)
 
         model = tf.keras.models.Model(
             inputs=self.image_input, outputs=out, name="binary_alexnet"
